Version 2.0
v2.0 includes a new function in adafruit_data.py to get the feed names from adafruit.io and add them as dictionary keys. This means if we add a new feed in the future we don't have to specifically call for the new feed.
Also added a set of if-elif-else statements in OpenWeather_data.py to compare the dictionary keys to variable names and add the corresponding values.